move GDALConfigManager to hyp3_gamma.util
