+++ hyp3_gamma/util.py
@@ -10,6 +10,28 @@ log = logging.getLogger(__name__)
 gdal.UseExceptions()
 
 
+class GDALConfigManager:
+    """Context manager for setting GDAL config options temporarily"""
+    def __init__(self, **options):
+        """
+        Args:
+            **options: GDAL Config `option=value` keyword arguments.
+        """
+        self.options = options.copy()
+        self._previous_options = {}
+
+    def __enter__(self):
+        for key in self.options:
+            self._previous_options[key] = gdal.GetConfigOption(key)
+
+        for key, value in self.options.items():
+            gdal.SetConfigOption(key, value)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        for key, value in self._previous_options.items():
+            gdal.SetConfigOption(key, value)
+
+
 def get_granule(granule):
     download_url = get_download_url(granule)
     zip_file = download_file(download_url, chunk_size=10485760)
