+++ tests/test_dem.py
@@ -9,29 +9,6 @@ from hyp3_gamma import dem
 gdal.UseExceptions()
 
 
-def test_gdal_config_manager():
-    gdal.SetConfigOption('OPTION1', 'VALUE1')
-    gdal.SetConfigOption('OPTION2', 'VALUE2')
-
-    assert gdal.GetConfigOption('OPTION1') == 'VALUE1'
-    assert gdal.GetConfigOption('OPTION2') == 'VALUE2'
-    assert gdal.GetConfigOption('OPTION3') is None
-    assert gdal.GetConfigOption('OPTION4') is None
-
-    with dem.GDALConfigManager(OPTION2='CHANGED', OPTION3='VALUE3'):
-        assert gdal.GetConfigOption('OPTION1') == 'VALUE1'
-        assert gdal.GetConfigOption('OPTION2') == 'CHANGED'
-        assert gdal.GetConfigOption('OPTION3') == 'VALUE3'
-        assert gdal.GetConfigOption('OPTION4') is None
-
-        gdal.SetConfigOption('OPTION4', 'VALUE4')
-
-    assert gdal.GetConfigOption('OPTION1') == 'VALUE1'
-    assert gdal.GetConfigOption('OPTION2') == 'VALUE2'
-    assert gdal.GetConfigOption('OPTION3') is None
-    assert gdal.GetConfigOption('OPTION4') == 'VALUE4'
-
-
 def test_get_geometry_from_kml(test_data_dir):
     kml = test_data_dir / 'alaska.kml'
     expected = {
